use u32 for config dimensions

+++ crates/backend-uzu/src/config/decoder.rs
@@ -9,6 +9,6 @@ pub struct DecoderConfig {
     pub embedding_config: AnyEmbeddingConfig,
     pub transformer_config: TransformerConfig,
 
-    pub vocab_size: usize,
+    pub vocab_size: u32,
     pub ple_model_config: Option<PLEModelConfig>,
 }
