+++ crates/backend-uzu/src/config/weight_matrix/mlx_spec.rs
@@ -5,6 +5,6 @@ use crate::config::weight_matrix::Layout;
 #[uzu_config(super::WeightMatrixSpec)]
 pub struct MLXSpec {
     pub bits: u32,
-    pub group_size: usize,
+    pub group_size: u32,
     pub layout: Layout,
 }
